skip existence check
Image nodes jump past the content of their alt text, which means there is no chance that `addNode` will be called when the last node in `blocks` is an image. Since images are the only node type pushed to `blocks` that has no `block` property, we can remove the existence check in the `addNode` function.